# api/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "TennisProject",
      "version": "1.0"
    },
    "paths": {
      "/api/Auth/register": {
        "post": {
          "tags": [
            "Auth"
          ],
          "requestBody": {
            "content": {
              "application/json-patch+json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRegisterRequest"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRegisterRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRegisterRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRegisterRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/Auth/initDataValidate": {
        "post": {
          "tags": [
            "Auth"
          ],
          "requestBody": {
            "content": {
              "application/json-patch+json": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              },
              "application/*+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/ValidatorResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ValidatorResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ValidatorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/User/profile": {
        "get": {
          "tags": [
            "User"
          ],
          "requestBody": {
            "content": {
              "application/json-patch+json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserRequest"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "AuthRegisterRequest": {
          "type": "object",
          "properties": {
            "user": {
              "$ref": "#/components/schemas/User"
            }
          },
          "additionalProperties": false
        },
        "GetUserRequest": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            },
            "telegramId": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "telegramId": {
              "type": "integer",
              "format": "int64"
            },
            "telegramUsername": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "lastName": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            },
            "sex": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "interestsId": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "occupation": {
              "type": "string",
              "nullable": true
            },
            "accountStatus": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "tennisLevel": {
              "type": "number",
              "format": "double",
              "nullable": true
            },
            "registrationDate": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "ValidatorResponse": {
          "type": "object",
          "properties": {
            "isValidated": {
              "type": "boolean"
            },
            "telegramId": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            },
            "message": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        }
      }
    }
  }